heathbot/lambda/bot_post/lambda-requirements.txt
discord.py==2.2.2
pytz==2022.7.1
comics==0.9.1
boto3